% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msteams.R
\name{make_base_card}
\alias{make_base_card}
\title{Make card wrapper}
\usage{
make_base_card(task_name, status, items = NULL)
}
\arguments{
\item{task_name}{Task name}

\item{status}{Task status}

\item{items}{List of objects to go into a card body (see https://adaptivecards.microsoft.com/designer.html)}
}
\description{
Creates message and card wrappers around the content of a card. Designed to
be part of a card creator.
}
