% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msteams.R
\name{make_columnset}
\alias{make_columnset}
\title{Make columnset}
\usage{
make_columnset(targ_df, cols)
}
\arguments{
\item{targ_df}{Dataframe to generate the columnset from}

\item{cols}{Names of the columns to use}
}
\description{
Turn dataframe columns into a columnset for AdapativeCards.
}
