% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msteams.R
\name{make_error_block}
\alias{make_error_block}
\title{Make error block}
\usage{
make_error_block(err_msg)
}
\arguments{
\item{err_msg}{Error message (only the $message component of the error!)}
}
\description{
Turn error messages into a text block for AdapativeCards.
}
