% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{run}
\alias{run}
\title{One-line run}
\usage{
run(project_name, upload_targets, container_url, maintainers)
}
\arguments{
\item{project_name}{Project name}

\item{upload_targets}{Vector of name-strings for targets that should be uploaded}

\item{container_url}{Azure container URL}

\item{maintainers}{List of maintainers to ping}
}
\description{
Runs targets, saves data locally, stores data on the blob
}
