% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{run_targets}
\alias{run_targets}
\title{Wrapper for running targets}
\usage{
run_targets(run_name, project_name, ping = c())
}
\arguments{
\item{run_name}{Run name}

\item{project_name}{Project name}

\item{ping}{Ping users in this message using their emails (case sensitive) as identifiers}
}
\description{
Wrapper for automated target runs. Runs tar_make() and sends a Teams
message, even on failures.
}
