% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{save_run_data}
\alias{save_run_data}
\title{Save files from current targets run}
\usage{
save_run_data(run_name, output_path = "outputs", curr_name = "current")
}
\arguments{
\item{run_name}{Run name}

\item{output_path}{Base output folder}
}
\description{
Saves copies files from target folder into a named folder
}
