% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msteams.R
\name{send_msg}
\alias{send_msg}
\title{Send a simple message}
\usage{
send_msg(msg, ...)
}
\arguments{
\item{msg}{Message to send}

\item{ping}{Ping users in this message using their emails (case sensitive) as identifiers}

\item{summary}{Summary message that'll be displayed on previews}
}
\description{
Sends a simple Teams message inside an AdaptiveCards using the Workflow
webhook to a specific channel.
}
