% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{send_run_report}
\alias{send_run_report}
\title{Send run report}
\usage{
send_run_report(run_name, project_name, ping, err_msg = NULL)
}
\arguments{
\item{run_name}{Run name}

\item{project_name}{Project name}

\item{ping}{Ping users in this message using their emails (case sensitive) as identifiers}

\item{err_msg}{Error message ($message component of error object)}
}
\description{
Generate and send a fancy formatted Teams message describing the outcome of
a targets run.
}
