% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msteams.R
\name{send_card}
\alias{send_card}
\title{Underlying function for sending cards}
\usage{
send_card(body, ping = NULL, summary = "")
}
\arguments{
\item{body}{Body of AdaptiveCard (list of AdaptiveCard elements expected)}

\item{ping}{Ping users in this message using their emails (case sensitive) as identifiers}

\item{summary}{Summary message that'll be displayed on previews}
}
\description{
Sends Teams AdaptiveCards using the Workflow webhook to a specific channel.
The webhook should expect to receive messages containing AdaptiveCard
attachments.
}
\details{
The payload needs to be in JSON, but we are generating JSON-like structures
in R (using list() to describe dictionaries and arrays) and then converting
it into a JSON string. The content also needs to follow very specific rules.
See make_card_payload() to understand how to attach an AdaptiveCard to a
message, and use the WYSIWYG designer to understand how to create an
AdaptiveCard: https://adaptivecards.microsoft.com/designer.html

If you run into trouble, test with a JSON string that is known to work and
step up from there, e.g.:
'\link{{ "type": "TextBlock", "text": "Hello World!" }}'

In R JSON-like structure, this is:
list(list(type = "TextBlock", text = "Hello World!"))
}
