% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{send_run_report}
\alias{send_run_report}
\title{Send run report}
\usage{
send_run_report(project_name, run_name, ping)
}
\arguments{
\item{project_name}{Project name}

\item{run_name}{Run name}

\item{ping}{Ping users in this message using their emails (case sensitive) as identifiers}
}
\description{
Generate and send a fancy formatted Teams message describing the outcome of
a targets run.
}
