% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{store_run_data}
\alias{store_run_data}
\title{Store files from targets run}
\usage{
store_run_data(
  run_name,
  project_name,
  container_url,
  upload_targets = c(),
  upload_folders = c(),
  update = TRUE,
  forced = FALSE
)
}
\arguments{
\item{run_name}{Run name}

\item{project_name}{Project name}

\item{container_url}{Azure container URL}

\item{upload_targets}{Vector of name-strings for targets that should be uploaded}

\item{upload_folders}{Vector of name-strings for folders that should be uploaded}

\item{forced}{Overwrite blob version}
}
\description{
Stores a specific list of upload targets, validation files and metadata.
}
